refactored the file structure of the frontend components
