+++ frontend/src/components/event/EventDisplay.jsx
@@ -1,11 +1,11 @@
 
-import styles from '../styles/Landing.module.css'
+import styles from '../../styles/Landing.module.css'
 import { useEffect, useState } from 'react';
-import EventTableEntry from '../components/EventTableEntry'
+import EventTableEntry from './EventTableEntry.jsx'
 
-import userService from '../services/userService';
-import eventService from '../services/eventService';
-import EventDetailsParramTitles from '../components/EventDetailsParramTitles';
+import userService from '../../services/userService.js';
+import eventService from '../../services/eventService.js';
+import EventDetailsParramTitles from './EventDetailsParramTitles.jsx';
 
 
 const EventDisplay = ({ user }) => {
